Issues fixed: #50: check that reverse dns resolves back to client_ip #51: checks against HELO
Two new checks, reverse and helo. Behind these is a 
completely new dns helper mechanism. The idea is to let a
single thread to handle all dns queries. Eventually it
should lead to fewer idling threads which is a good thing
on some platforms.

Now it needs to be tested on different platforms.